Move query keyword quoting into make_query
